The actual game file
A starter game I wrote when I first started making personal projects, though not a lot at the time it was a start and I hope yall enjoy it!!! Snake Game 🐍 is a simple arcade game built with Python and Pygame. Guide the snake to eat food, grow longer, and avoid collisions with walls or yourself while chasing the highest score.